+++ tedi/components/buttons/index.ts
@@ -1,6 +1,7 @@
 export * from "./button/button.component";
 export * from "./button-group/button-group.component";
 export * from "./button-group/button-group-button/button-group-button.directive";
+export * from "./card-button";
 export * from "./info-button/info-button.component";
 export * from "./collapse/collapse.component";
 export * from "./collapse-button/collapse-button.component";